tokenizer: Store keywords in list instead of map
At some point recently the Odin compiler disabled the ability to have
file-scope dynamic type literals by default and need a special feature
flag to do that. Rather than enabling that feature, I just changed it to
be a linear seach over an array which is probably just as good given the
small amount of keywords to look through.